requirements.txt
playwright>=1.40.0
pillow>=10.0.0
feedparser>=6.0.10
